# Conflicts:
#	back-end/src/main/java/com/petmily/backend/support/donate/controller/DonationController.java
#	back-end/src/main/java/com/petmily/backend/support/donate/service/DonationService.java
#	back-end/src/main/java/com/petmily/backend/support/volunteer/controller/VolunteerController.java
#	back-end/src/main/java/com/petmily/backend/support/volunteer/repository/VolunteerRepository.java
#	back-end/src/main/java/com/petmily/backend/support/volunteer/service/VolunteerService.java
#	back-end/src/main/java/com/petmily/backend/support/volunteerReview/controller/VolunteerReviewController.java
#	back-end/src/main/java/com/petmily/backend/support/volunteerReview/repository/VolunteerReviewRepository.java
#	back-end/src/main/java/com/petmily/backend/support/volunteerReview/service/VolunteerReviewService.java

+++ back-end/src/main/java/com/petmily/backend/comment/dto/CommentDto.java
@@ -15,7 +15,7 @@ public class CommentDto {
     private String commentContent;
     private LocalDateTime commentCreate;
     private LocalDateTime commentUpdate;
-    private Integer commentPnum;
+    private Long commentPnum;
     private Boolean commentIsSecret;
 
     private String memberImg;
